+++ packages/sdk-nextjs/lib/workflow-engine/core/handlers/text-generation.ts
