+++ packages/sdk-nextjs/package.json
@@ -14,6 +14,7 @@
 		"next": "15.1.4",
 		"react": "19.0.0",
 		"react-dom": "19.0.0",
+		"unstorage": "1.14.4",
 		"zod": "3.24.1"
 	},
 	"devDependencies": {
